+++ tests/Feature/Components/Alert/AlertTest.php
@@ -1,0 +1,70 @@
+<?php
+
+namespace Bjnstnkvc\ShadcnUi\Tests\Feature\Components\Alert;
+
+use Bjnstnkvc\ShadcnUi\Tests\TestCase;
+use PHPUnit\Framework\Attributes\Test;
+
+class AlertTest extends TestCase
+{
+    #[Test]
+    public function it_renders_the_component()
+    {
+        $template = <<<'HTML'
+            <x-alert></x-alert>
+        HTML;
+
+        $this->assertStringContainsString(
+            needle  : '<div class="relative w-full rounded-lg border [&>svg~*]:pl-7 [&>svg+div]:translate-y-[-3px] [&>svg]:absolute [&>svg]:left-4 [&>svg]:top-4 [&>svg]:text-foreground p-4 bg-background text-foreground" role="alert">',
+            haystack: $this->minify($this->blade($template)),
+        );
+    }
+
+    #[Test]
+    public function it_renders_the_component_with_props()
+    {
+        $template = <<<'HTML'
+            <x-alert variant="destructive"></x-alert>
+        HTML;
+
+        $this->assertStringContainsString(
+            needle  : '<div class="relative w-full rounded-lg border [&>svg~*]:pl-7 [&>svg+div]:translate-y-[-3px] [&>svg]:absolute [&>svg]:left-4 [&>svg]:top-4 [&>svg]:text-foreground p-4 border-destructive/50 text-destructive [&>svg]:text-destructive" role="alert">',
+            haystack: $this->minify($this->blade($template)),
+        );
+    }
+
+    #[Test]
+    public function it_renders_the_component_with_slot()
+    {
+        $template = <<<'HTML'
+            <x-alert>
+                <div>Hello World</div>
+            </x-alert>
+        HTML;
+
+        $this->assertStringContainsString(
+            needle  : '<div>Hello World</div>',
+            haystack: $this->minify($this->blade($template)),
+        );
+    }
+
+    #[Test]
+    public function it_renders_the_component_as_child()
+    {
+        $template = <<<'HTML'
+            <x-alert as-child>
+                <section></section>
+            </x-alert>
+        HTML;
+
+        $this->assertStringNotContainsString(
+            needle  : '<div class="relative w-full rounded-lg border [&>svg~*]:pl-7 [&>svg+div]:translate-y-[-3px] [&>svg]:absolute [&>svg]:left-4 [&>svg]:top-4 [&>svg]:text-foreground p-4 bg-background text-foreground" role="alert">',
+            haystack: $this->minify($this->blade($template)),
+        );
+
+        $this->assertStringContainsString(
+            needle  : '<section class="relative w-full rounded-lg border [&>svg~*]:pl-7 [&>svg+div]:translate-y-[-3px] [&>svg]:absolute [&>svg]:left-4 [&>svg]:top-4 [&>svg]:text-foreground p-4 bg-background text-foreground" role="alert">',
+            haystack: $this->minify($this->blade($template)),
+        );
+    }
+}