+++ CHANGELOG.md
@@ -4,5 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
-## 1.0.0 (2025-01-03)
+## 1.0.1 (2025-01-03)
+
+### Added
+
+- Added Alert component.
+
+### Changed
+
+- Added `app.css` content to Tailwind section in `README.md`.
+- Changed snipped template from `bladehtml` to `html`.
+
+## 1.0.0 (2025-01-02)
+
 Initial release.